被弾処理作成、軽度修正

+++ SpriteComponent.cpp
@@ -94,7 +94,7 @@ void SpriteComponent::Draw() {
 
     renderer->DrawSpriteEx(
         pos, scale.x, scale.y, 0.0f, m_handle,
-        true, Vector2d(texW, texH) * 0.5f, 255, false, false
+        true, Vector2d((float)texW, (float)texH) * 0.5f, 255, false, false
     );
 }
